Add files via upload
Because there are different ways to convert types in java, it is a little bit clunky and inconvenient. I made this intuitive class for converting strings, integers, and booleans. I may add char, byte, short, long, doubles, and floats in the future. Attached are the version with comments explaining the code and a compact version to download or copy and paste. This is my first contribution and I am relatively new to java so I hope you enjoy!